+++ lib/Navigation/Iterator/PrefixRecursiveFilterIterator.php
@@ -1,16 +1,38 @@
 <?php
+
 declare(strict_types=1);
 
-namespace Pimcore\Navigation;
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Commercial License (PCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PCL
+ */
+
+namespace Pimcore\Navigation\Iterator;
+
+use Pimcore\Navigation\Container;
+use Pimcore\Navigation\Page;
 
 /**
  * @internal
  */
-final class ContainerRecursiveFilterIterator extends \RecursiveFilterIterator
+final class PrefixRecursiveFilterIterator extends \RecursiveFilterIterator
 {
     private string $property;
     private string $value;
 
+    /**
+     * @param Container $iterator navigation container to iterate
+     * @param string $property name of property that acts as needle
+     * @param string $value value which acts as haystack
+     */
     public function __construct(Container $iterator, string $property, string $value)
     {
         parent::__construct($iterator);
